# src/main/resources/api/service_openapi.yaml
openapi: 3.0.0
info:
  title: Folder Manager API
  description: |
    Folder Manager REST API.
  version: 0.0.1

servers:
  - url: http://localhost:8080

paths:
  '/status':
    get:
      description: |
        Returns the operational status of the service
      operationId: serviceStatus
      tags:
        - unauthenticated
      responses:
        200:
          description: Service is functional
          $ref: '#/components/responses/StatusResponse'
        500:
          description: Service is broken
          $ref: '#/components/responses/StatusResponse'
  '/api/v1/folders':
    post:
      description: |
        Synchronous endpoint to create a new empty folder.
      operationId: createFolder
      tags:
      - folder
      requestBody:
        description: |
          Name of the new folder to be created, as well as the ID of the
          containing folder (absent if this is a top-level folder).
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderBody'
      responses:
        200:
          description: Ok
          $ref: '#/components/responses/CreatedFolderResponse'
        400:
          description: Bad request - invalid id, badly formed
          $ref: '#/components/responses/ErrorResponse'
        403:
          description: Permission denied
          $ref: '#/components/responses/ErrorResponse'
  '/api/v1/jobs/{id}':
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      description: |
        Poll the status of an existing async operation via job ID.
      operationId: pollAsyncJob
      tags:
      - folder
      responses:
        202:
          description: Async job is incomplete
          $ref: '#/components/responses/JobModelResponse'
        200:
          description: Async job is complete
          $ref: '#/components/responses/JobModelResponse'
        400:
          description: Bad request - invalid id, badly formed
          $ref: '#/components/responses/ErrorResponse'
        403:
          description: No permission to see job
          $ref: '#/components/responses/ErrorResponse'
        404:
          description: Not found - job id does not exist
          $ref: '#/components/responses/ErrorResponse'
    delete:  ## delete job data
      description: Delete the job and data associated with it
      operationId: deleteJob
      tags:
      - folder
      responses:
        204:
          description: Job was deleted
  '/api/v1/jobs/{id}/result':
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      description: |
        Retrieve the results of a completed async job specified by ID.
      operationId: retrieveJobResult
      tags:
      - folder
      responses:
        default:
          description: Successful responses return the type of object specified by the job; otherwise, ErrorModel
          content:
            application/json:
              schema:
                type: object
components:
  parameters:
    Id:
      name: id
      in: path
      description: A UUID to used to identify an object in the folder manager
      required: true
      schema:
        type: string
  schemas:
    ErrorReport:
      type: object
      properties:
        message:
          type: string
        statusCode:
          type: integer
        causes:
          type: array
          items:
            type: string
    JobControl:
      type: object
      required:
      - jobid
      properties:
        jobid:
          type: string
        pubsub:
          type: object
          required:
          - projectid
          - topicid
          properties:
            projectid:
              type: string
            topicid:
              type: string
    JobModel:
      type: object
      required:
      - id
      - status
      - status_code
      properties:
        id:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - RUNNING
          - SUCCEEDED
          - FAILED
        statusCode:
          description: HTTP code
          type: integer
        submitted:
          type: string
        completed:
          type: string
        estimatedRemainingTimeMS:
          type: integer
    SystemStatus:
      type: object
      properties:
        ok:
          type: boolean
          description: status of this service
        systems:
          type: object
          additionalProperties:
            type: object
            properties:
              ok:
                type: boolean
              messages:
                type: array
                items:
                  type: string
    CreateFolderBody:
      type: object
      required:
      - name
      - jobControl
      properties:
        name:
          type: string
          description: Name of the new folder to create
        parentFolderId:
          type: string
          nullable: true
          description: |
            UUID of the parent folder to contain the newly-created folder.
        spendProfile:
          type: string
          nullable: true
          description: |
            Spend profile ID to use for this folder. Cannot be set if the parent
            folder already provides a spend profile.
    CreatedFolder:
      type: object
      properties:
        id:
          type: string
          description: UUID of a newly-created Folder.

  responses:
    ErrorResponse:
      description: common error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReport'

    StatusResponse:
      description: common status response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SystemStatus'

    JobModelResponse:
      description: Response with a JobModel
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JobModel'
    CreatedFolderResponse:
      description: Response with a CreatedFolder
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreatedFolder'